requirements.txt
aiogram==2.14.3
aiohttp==3.7.4.post0
altgraph==0.17
async-timeout==3.0.1
attrs==21.2.0
Babel==2.9.1
certifi==2021.5.30
chardet==4.0.0
future==0.18.2
idna==3.2
importlib-metadata==4.8.0
multidict==5.1.0
numpy==1.19.4
pandas==1.1.5
pefile==2021.5.24
pyaes==1.6.1
pyasn1==0.4.8
pyinstaller==4.5.1
pyinstaller-hooks-contrib==2021.3
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2020.5
pywin32-ctypes==0.2.0
rsa==4.7.2
selenium==3.141.0
six==1.15.0
Telethon==1.23.0
typing-extensions==3.10.0.0
urllib3==1.26.2
yarl==1.6.3
zipp==3.5.0
